agregar clases

+++ Lab2_Alberto Martinez/src/lab2_alberto/martinez/Clase.java
@@ -17,10 +17,10 @@ public class Clase {
     public int max;
     public Maestro maestro;
     public int unidades;
-    public int precio;
+    public double precio;
     public ArrayList <Alumno>alumnos=new ArrayList();
 
-    public Clase(String nombre, int seccion, int max, Maestro maestro, int unidades, int precio) {
+    public Clase(String nombre, int seccion, int max, Maestro maestro, int unidades, double precio) {
         this.nombre = nombre;
         this.seccion = seccion;
         this.max = max;
@@ -79,7 +79,7 @@ public class Clase {
         this.alumnos = alumnos;
     }
 
-    public int getPrecio() {
+    public double getPrecio() {
         return precio;
     }
